+++ CHANGELOG.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added BART model
 - Added `ModelCard` and related classes. Added model cards for all the pretrained models.
-- Added a field `registered_predictor_name` to `ModelCard` and a method named `load_predictor`.
+- Added a field `registered_predictor_name` to `ModelCard`.
+- Added a method `load_predictor` to `allennlp_models.pretrained`.
 
 ## [v1.1.0rc1](https://github.com/allenai/allennlp-models/releases/tag/v1.1.0rc1) - 2020-07-14